requirements.txt
pandas
pyarrow
spotipy
python-dotenv
tqdm
pyyaml
requests
tqdm